feat: refactors webshot service as a lib, adds missing endpoint to scenario-service
