+++ api/libs/webshot/src/index.ts
@@ -1,0 +1,4 @@
+export * from './webshot.module';
+export { WebshotService, unknownPdfWebshotError } from './webshot.service';
+
+export { WebshotSummaryReportConfig } from './webshot.dto';
