+++ api/libs/webshot/src/webshot.dto.ts
@@ -1,0 +1,17 @@
+import { IsOptional, IsString } from 'class-validator';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { PDFOptions } from 'puppeteer';
+
+export class WebshotSummaryReportConfig {
+  @ApiProperty()
+  @IsString()
+  baseUrl!: string;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  cookie?: string;
+
+  @ApiPropertyOptional()
+  @IsOptional()
+  pdfOptions?: PDFOptions;
+}
