+++ api/libs/webshot/src/webshot.module.ts
@@ -1,0 +1,10 @@
+import { HttpModule, Module } from '@nestjs/common';
+import { WebshotService } from './webshot.service';
+
+@Module({
+  imports: [HttpModule],
+  providers: [WebshotService],
+  controllers: [],
+  exports: [WebshotService],
+})
+export class WebshotModule {}
